+++ docs-demo/basic/multi-header/MultiHeader.vue
@@ -1,8 +1,12 @@
 <script setup lang="ts">
+import { ref } from 'vue';
 import StkTable from '../../StkTable.vue';
+import CheckItem from '../../components/CheckItem.vue';
 import { StkTableColumn } from '../../../src/StkTable/index';
 import { useI18n } from '../../hooks/useI18n';
+
 const { t } = useI18n();
+const virtual = ref(true);
 const columns: StkTableColumn<any>[] = [
     {
         dataIndex: 'Basic',
@@ -37,7 +41,7 @@ const columns: StkTableColumn<any>[] = [
     {
         dataIndex: 'other',
         title: t('other'),
-        children: new Array(2).fill(0).map((it, i) => {
+        children: new Array(4).fill(0).map((it, i) => {
             return {
                 dataIndex: 'other' + i,
                 title: t('other') + ' ' + i,
@@ -60,16 +64,17 @@ const dataSource = new Array(50).fill(0).map((it, i) => {
         lv2_1: 'lv2.1',
         lv2_2: 'lv2.2',
         age: i,
-        email: i + '@qq.com',
+        email: i + '@email.com',
     };
 });
 </script>
 <template>
+    <CheckItem v-model="virtual" text="virtual"></CheckItem>
     <StkTable
         style="height: 200px"
         row-key="id"
         fixed-col-shadow
-        virtual
+        :virtual="virtual"
         :columns="columns"
         :data-source="dataSource"
     >