+++ docs-demo/basic/multi-header/MultiHeaderAnyFixed.vue
@@ -66,14 +66,12 @@ const dataSource = new Array(50).fill(0).map((it, i) => {
         lv2_1: 'lv2.1',
         lv2_2: 'lv2.2',
         age: i,
-        email: i + '@qq.com',
+        email: i + '@email.com',
     };
 });
 
 onMounted(() => {
-    window.setTimeout(() => {
-        (stkTableRef.value as any)?.scrollTo(0, 100);
-    });
+    stkTableRef.value?.scrollTo(0, 100);
 });
 </script>
 <template>