+++ docs-src/.vitepress/src/config/en.ts
@@ -55,14 +55,14 @@ export const enConfig = defineConfig({
                                     { text: 'Headless', link: '/table/basic/headless' },
                                     { text: 'Row Expansion', link: '/table/basic/expand-row' },
                                     { text: 'Tree', link: '/table/basic/tree' },
-                                    { text: 'Multi-level Header', link: '/table/basic/multi-header' },
+                                    { text: 'Multi-level Header(✨Update)', link: '/table/basic/multi-header' },
                                     { text: 'Sequence Column', link: '/table/basic/seq' },
                                     { text: 'Empty Data', link: '/table/basic/empty' },
                                     { text: 'Row & Column Unique Keys', link: '/table/basic/key' },
                                     { text: 'Scrollbar', link: '/table/basic/scrollbar' },
                                     { text: 'Table-layout: fixed', link: '/table/basic/fixed-mode' },
                                     { text: 'Row-by-Row Scrolling', link: '/table/basic/scroll-row-by-row' },
-                                    { text: 'Footer(✨NEW)', link: '/table/basic/footer' },
+                                    { text: 'Footer', link: '/table/basic/footer' },
                                 ]
                             },
                             {
@@ -79,6 +79,15 @@ export const enConfig = defineConfig({
                                     { text: 'Custom Cell', link: '/table/advanced/custom-cell' },
                                     { text: 'Custom Sorting', link: '/table/advanced/custom-sort' },
                                     { text: 'Vue 2 Scroll Optimization', link: '/table/advanced/vue2-scroll-optimize' },
+                                    {
+                                        text: 'Built-in Custom Cell Extensions (✨NEW)',
+                                        collapsed: true,
+                                        items: [
+                                            { text: 'EditableCell', link: '/table/advanced/custom-cells/editable-cell' },
+                                            { text: 'FilterCell', link: '/table/advanced/custom-cells/filter-cell' },
+                                            { text: 'CheckboxCell', link: '/table/advanced/custom-cells/checkbox-cell' },
+                                        ],
+                                    },
                                 ]
                             },
                             {