+++ docs-src/.vitepress/src/config/ko.ts
@@ -55,14 +55,14 @@ export const koConfig = defineConfig({
                                     { text: '헤더리스', link: '/table/basic/headless' },
                                     { text: '행 확장', link: '/table/basic/expand-row' },
                                     { text: '트리', link: '/table/basic/tree' },
-                                    { text: '멀티 레벨 헤더', link: '/table/basic/multi-header' },
+                                    { text: '멀티 레벨 헤더(✨Update)', link: '/table/basic/multi-header' },
                                     { text: '시퀀스 열', link: '/table/basic/seq' },
                                     { text: '빈 데이터', link: '/table/basic/empty' },
                                     { text: '행/열 고유 키', link: '/table/basic/key' },
                                     { text: '스크롤바', link: '/table/basic/scrollbar' },
                                     { text: 'table-layout: fixed', link: '/table/basic/fixed-mode' },
                                     { text: '행 단위 스크롤', link: '/table/basic/scroll-row-by-row' },
-                                    { text: '푸터 (✨NEW)', link: '/table/basic/footer' },
+                                    { text: '푸터', link: '/table/basic/footer' },
                                 ]
                             },
                             {
@@ -79,6 +79,15 @@ export const koConfig = defineConfig({
                                     { text: '커스텀 셀', link: '/table/advanced/custom-cell' },
                                     { text: '커스텀 정렬', link: '/table/advanced/custom-sort' },
                                     { text: 'Vue 2 스크롤 최적화', link: '/table/advanced/vue2-scroll-optimize' },
+                                    {
+                                        text: '커스텀 셀 내장 확장(✨NEW)',
+                                        collapsed: true,
+                                        items: [
+                                            { text: 'EditableCell 편집 가능 셀', link: '/table/advanced/custom-cells/editable-cell' },
+                                            { text: 'FilterCell 필터', link: '/table/advanced/custom-cells/filter-cell' },
+                                            { text: 'CheckboxCell 체크박스', link: '/table/advanced/custom-cells/checkbox-cell' },
+                                        ],
+                                    },
                                 ]
                             },
                             {