+++ lib/src/StkTable/custom-cells/Filter/Filter.vue.d.ts
@@ -4,19 +4,13 @@ import { FilterOption } from './types';
 declare const _default: import('vue').DefineComponent<import('vue').ExtractPropTypes<__VLS_TypePropsToRuntimeProps<CustomHeaderCellProps<any> & {
     theme?: "light" | "dark";
     active?: boolean;
-    options?: FilterOption[];
-    dataSource: any[];
-    col: StkTableColumn<any>;
-    colIndex: number;
+    getOptions: () => FilterOption[];
 }>>, {}, {}, {}, {}, import('vue').ComponentOptionsMixin, import('vue').ComponentOptionsMixin, {
     change: (value: any[], column: StkTableColumn<any>, colIndex: number) => void;
 }, string, import('vue').PublicProps, Readonly<import('vue').ExtractPropTypes<__VLS_TypePropsToRuntimeProps<CustomHeaderCellProps<any> & {
     theme?: "light" | "dark";
     active?: boolean;
-    options?: FilterOption[];
-    dataSource: any[];
-    col: StkTableColumn<any>;
-    colIndex: number;
+    getOptions: () => FilterOption[];
 }>>> & Readonly<{
     onChange?: ((value: any[], column: StkTableColumn<any>, colIndex: number) => any) | undefined;
 }>, {}, {}, {}, {}, string, import('vue').ComponentProvideOptions, true, {}, any>;