+++ lib/src/StkTable/custom-cells/Filter/types.d.ts
@@ -1,4 +1,4 @@
-import { StkTableColumn } from '../../types';
+import { UniqKey } from '../../types';
 
 /** 筛选选项接口 */
 export interface FilterOption {
@@ -39,4 +39,9 @@ export interface FilterComponentConfig {
 export interface CreateFilterOption {
     /** 是否远程筛选 */
     remote?: boolean;
+    /** 筛选状态改变时触发 */
+    onChange?: (data: {
+        colKey: UniqKey;
+        status: FilterStatus;
+    }) => void;
 }