+++ src/StkTable/StkTable.vue
@@ -62,14 +62,18 @@
                 @mouseover="onTrMouseOver"
             >
                 <thead v-if="!headless">
-                    <tr v-for="(row, rowIndex) in tableHeaders" :key="rowIndex" @contextmenu="onHeaderMenu($event)">
+                    <tr
+                        v-for="(row, rowIndex) in virtualX_on ? virtualX_tableHeaders : tableHeaders"
+                        :key="rowIndex"
+                        @contextmenu="onHeaderMenu($event)"
+                    >
                         <th
                             v-if="virtualX_on"
                             class="vt-x-left"
-                            :style="`min-width:${virtualScrollX.offsetLeft}px;width:${virtualScrollX.offsetLeft}px`"
+                            :style="`min-width:${theadVirtualX.offsetLeft}px;width:${theadVirtualX.offsetLeft}px`"
                         ></th>
                         <th
-                            v-for="(col, colIndex) in virtualX_on && rowIndex === tableHeaders.length - 1 ? virtualX_columnPart : row"
+                            v-for="(col, colIndex) in row"
                             :key="colKeyGen(col)"
                             v-bind="getTHProps(col)"
                             @click="e => onHeaderCellClick(e, col)"
@@ -82,7 +86,7 @@
                                 class="table-header-resizer left"
                                 @mousedown="onThResizeMouseDown($event, col, true)"
                             ></div>
-                            <div class="table-header-cell-wrapper" :style="virtualX_on ? null : col.__R_SP__ ? `--row-span:${col.__R_SP__}` : null">
+                            <div class="table-header-cell-wrapper" :style="col.__R_SP__ ? `--row-span:${col.__R_SP__}` : null">
                                 <component
                                     :is="col.customHeaderCell"
                                     v-if="col.customHeaderCell"
@@ -126,24 +130,32 @@
                     :style="isFooterTop ? `top:${tableHeaderHeight}px` : ''"
                 >
                     <tr v-for="(footRow, footRowIndex) in footerData" :key="footRowIndex">
-                        <td v-if="virtualX_on" class="vt-x-left"></td>
-                        <td v-for="(col, colIndex) in virtualX_columnPart" :key="colKeyGen(col)" v-bind="getTFProps(col)">
-                            <component
-                                :is="col.customFooterCell"
-                                v-if="col.customFooterCell"
-                                class="table-cell-wrapper"
-                                tabindex="-1"
-                                :col="col"
-                                :row="footRow"
-                                :rowIndex="footRowIndex"
-                                :colIndex="colIndex"
-                                :cellValue="footRow[col.dataIndex]"
-                            />
-                            <div class="table-cell-wrapper" tabindex="-1" :title="footRow[col.dataIndex] || ''">
-                                <span v-if="footRow[col.dataIndex] != null">{{ footRow[col.dataIndex] }}</span>
-                            </div>
-                        </td>
-                        <td v-if="virtualX_on" class="vt-x-right"></td>
+                        <td
+                            v-if="virtualX_on"
+                            class="vt-x-left"
+                            :style="`min-width:${theadVirtualX.offsetLeft}px;width:${theadVirtualX.offsetLeft}px`"
+                        ></td>
+                        <template v-for="col in virtualX_columnPart" :key="col.type === 'spacer' ? 'spacer' : colKeyGen(col)">
+                            <template v-if="col.type === 'spacer'">
+                                <td v-if="col.__COLSPAN__" class="vt-x-spacer" :colspan="col.__COLSPAN__"></td>
+                            </template>
+                            <td v-else v-bind="getTFProps(col)">
+                                <component
+                                    :is="col.customFooterCell"
+                                    v-if="col.customFooterCell"
+                                    class="table-cell-wrapper"
+                                    tabindex="-1"
+                                    :col="col"
+                                    :row="footRow"
+                                    :rowIndex="footRowIndex"
+                                    :cellValue="footRow[col.dataIndex]"
+                                />
+                                <div class="table-cell-wrapper" tabindex="-1" :title="footRow[col.dataIndex] || ''">
+                                    <span v-if="footRow[col.dataIndex] != null">{{ footRow[col.dataIndex] }}</span>
+                                </div>
+                            </td>
+                        </template>
+                        <td v-if="virtualX_on" class="vt-x-right" :style="`min-width:${virtualX_offsetRight}px;width:${virtualX_offsetRight}px`"></td>
                     </tr>
                 </component>
 
@@ -157,14 +169,27 @@
                     @drop="onBodyDrop"
                 >
                     <tr v-if="!isExperimentalScrollY && virtual_on && !isSRBRActive" :style="paddingTopStyle" class="padding-top-tr">
-                        <td v-if="virtualX_on && fixedMode && headless" class="vt-x-left"></td>
                         <template v-if="fixedMode && headless">
-                            <td v-for="col in virtualX_columnPart" :key="colKeyGen(col)" :style="cellStyleMap[TagType.TD].get(colKeyGen(col))"></td>
+                            <td
+                                v-if="virtualX_on"
+                                class="vt-x-left"
+                                :style="`min-width:${theadVirtualX.offsetLeft}px;width:${theadVirtualX.offsetLeft}px`"
+                            ></td>
+                            <template v-for="col in virtualX_columnPart" :key="col.type === 'spacer' ? 'spacer' : colKeyGen(col)">
+                                <template v-if="col.type === 'spacer'">
+                                    <td v-if="col.__COLSPAN__" class="vt-x-spacer" :colspan="col.__COLSPAN__"></td>
+                                </template>
+                                <td v-else :style="cellStyleMap[TagType.TD].get(colKeyGen(col))"></td>
+                            </template>
+                            <td
+                                v-if="virtualX_on"
+                                class="vt-x-right"
+                                :style="`min-width:${virtualX_offsetRight}px;width:${virtualX_offsetRight}px`"
+                            ></td>
                         </template>
                     </tr>
                     <tr v-for="(row, rowIndex) in virtual_dataSourcePart" ref="trRef" :key="rowKeyGen(row)" v-bind="getTRProps(row, rowIndex)">
-                        <td v-if="virtualX_on" class="vt-x-left"></td>
-                        <td v-if="row && row.__EXP_R__" :colspan="virtualX_columnPart.length">
+                        <td v-if="row && row.__EXP_R__" :colspan="expandRowColspan">
                             <div class="table-cell-wrapper" tabindex="-1">
                                 <slot name="expand" :row="row.__EXP_R__" :col="row.__EXP_C__">
                                     {{ (row.__EXP_R__ && row.__EXP_C__ && row.__EXP_R__[row.__EXP_C__.dataIndex]) || '' }}
@@ -172,8 +197,15 @@
                             </div>
                         </td>
                         <template v-else>
-                            <template v-for="(col, colIndex) in virtualX_columnPart">
-                                <td v-if="!shouldHideCell(row, col)" :key="colKeyGen(col)" v-bind="getTDProps(row, col, rowIndex, colIndex)">
+                            <td v-if="virtualX_on" class="vt-x-left"></td>
+                            <template v-for="(col, colIndex) in virtualX_columnPart" :key="col.type === 'spacer' ? 'spacer' : colKeyGen(col)">
+                                <template v-if="col.type === 'spacer'">
+                                    <td v-if="col.__COLSPAN__" class="vt-x-spacer" :colspan="col.__COLSPAN__"></td>
+                                </template>
+                                <td
+                                    v-else-if="!shouldHideCell(row, col)"
+                                    v-bind="getTDProps(row, col, rowIndex, (col as PrivateStkTableColumn<DT>).__LEAF_START__ ?? 0)"
+                                >
                                     <template v-for="text in [getFormattedValue(row, col, rowIndex, colIndex)]" :key="text">
                                         <component
                                             :is="col.customCell"
@@ -222,8 +254,8 @@
                                     </template>
                                 </td>
                             </template>
+                            <td v-if="virtualX_on" class="vt-x-right"></td>
                         </template>
-                        <td v-if="virtualX_on" class="vt-x-right"></td>
                     </tr>
                     <template v-if="!isExperimentalScrollY">
                         <tr v-if="virtual_on && !isSRBRActive" :style="offsetBottomStyle"></tr>
@@ -278,7 +310,7 @@ import {
     DEFAULT_SORT_CONFIG,
     IS_LEGACY_MODE,
 } from './const';
-import type { FilterStatus } from './custom-cells/Filter/types';
+import type { FilterStatus } from './custom-cells/FilterCell/types';
 import { useAreaSelectionName } from './features';
 import { ON_DEMAND_FEATURE } from './registerFeature';
 import {
@@ -321,13 +353,13 @@ import { useTableColumns } from './useTableColumns';
 import { useThDrag } from './useThDrag';
 import { useTrDrag } from './useTrDrag';
 import { useTree } from './useTree';
-import Filter from './custom-cells/Filter/Filter.vue';
+import Filter from './custom-cells/FilterCell/Filter.vue';
 import { useIndexResolver } from './useIndexResolver';
 import { useVirtualScroll } from './useVirtualScroll';
 import { useWheeling } from './useWheeling';
 import { createStkTableId, getCalculatedColWidth } from './utils/constRefUtils';
 import { getClosestColKey, getClosestTd, getClosestTr, getClosestTrIndex, rafThrottle, transformWidthToStr } from './utils/index';
-import { FilterOption } from './custom-cells/Filter/types.js';
+import { FilterOption } from './custom-cells/FilterCell/types.js';
 
 /** Generic stands for DataType */
 type DT = any & PrivateRowDT;
@@ -901,7 +933,6 @@ const [
     virtual_dataSourcePart,
     virtual_offsetBottom,
     virtualX_on,
-    virtualX_columnPart,
     virtualX_offsetRight,
     tableHeaderHeight,
     initVirtualScroll,
@@ -912,6 +943,10 @@ const [
     setAutoHeight,
     clearAllAutoHeight,
     clearColWidthCache,
+    virtualX_tableHeaders,
+    expandRowColspan,
+    theadVirtualX,
+    virtualX_columnPart,
 ] = useVirtualScroll(
     props,
     tableContainerRef,
@@ -1155,6 +1190,7 @@ function setFilter(
     status: Record<UniqKey, FilterStatus | null> | null,
     option?: {
         remote?: boolean;
+        silent?: boolean;
     },
 ) {
     status = status || {};
@@ -1162,7 +1198,9 @@ function setFilter(
     if (!option?.remote) {
         initDataSource();
     }
-    emits('filter-change', status);
+    if (!option?.silent) {
+        emits('filter-change', status);
+    }
 }
 
 function filterDataSource(dataSource: DT[]) {
@@ -1268,8 +1306,13 @@ const cellStyleMap = computed(() => {
         for (let i = 0, colsLen = cols.length; i < colsLen; i++) {
             const col = cols[i];
             const width = virtualX ? getCalculatedColWidth(col) + 'px' : transformWidthToStr(col.width);
+            const minWidthStr = transformWidthToStr(col.minWidth);
+            const maxWidthStr = transformWidthToStr(col.maxWidth);
             // en: Use string instead of object to reduce patchStyle overhead
-            const styleStr = `--cw:${width};min-width:${transformWidthToStr(col.minWidth)};max-width:${transformWidthToStr(col.maxWidth)}`;
+            let styleStr = '';
+            if (width) styleStr += `--cw:${width}`;
+            if (minWidthStr) styleStr += `;min-width:${minWidthStr}`;
+            if (maxWidthStr) styleStr += `;max-width:${maxWidthStr}`;
             const colKey = colKeyGenValue(col);
             thMap.set(colKey, styleStr + ';' + getFixedStyle(TagType.TH, col, depth));
             tdMap.set(colKey, styleStr + ';' + getFixedStyle(TagType.TD, col, depth));
@@ -1343,7 +1386,7 @@ function getTHProps(col: PrivateStkTableColumn<DT>) {
     return {
         'data-col-key': colKey,
         draggable: Boolean(isHeaderDraggable(col)),
-        rowspan: virtualX_on.value ? 1 : col.__R_SP__,
+        rowspan: col.__R_SP__,
         colspan: col.__C_SP__,
         style: cellStyleMap.value[TagType.TH].get(colKey),
         title: getHeaderTitle(col),
@@ -1421,7 +1464,7 @@ function getTDProps(row: PrivateRowDT | null | undefined, col: StkTableColumn<Pr
         'data-col-key': colKey,
         style: cellStyleMap.value[TagType.TD].get(colKey),
         class: classList,
-        ...mergeCellsWrapper(row, col, rowIndex, colIndex),
+        ...mergeCellsWrapper(row, col, rowIndex, (col as PrivateStkTableColumn<DT>).__LEAF_START__ ?? 0),
     };
 }