+++ src/StkTable/custom-cells/Filter/index.ts
@@ -1,6 +1,0 @@
-// 导出Filter组件相关模块
-export { createFilter } from './createFilter';
-export type { FilterStatus, CreateFilterOption } from './types';
-// export { default as Filter } from './Filter.vue';
-
-import './Filter.less';